fix(graph/bfs): implement reasonable python queue
In Python, 'queue.Queue' is a 'multi-producer, multi-consumer queues. It is especially useful in threaded programming when information must be exchanged safely between multiple threads'.

If the 'Queue' is empty, and the programming is running, it’ll always waiting for `put`s, which may cause 'Time Limit Exceeds' in online coding platforms, such as Leetcode. 

Besides, using 'queue.Queue' is much slower than using `collections.deque` (which I changed it to) for the data structure `queue`.

P.S: Maybe code's not pythonic enough, because I tend to keep it, for beginners or people switched python to understand.